+++ NEWS.md
@@ -1,3 +1,17 @@
+# rjwsacruncher 0.2.3
+
+- add new parameter `standalone` to `download_cruncher()`.
+
+- A startup message is added indicating which version of the JWSACruncher is used by default (defined by the option `"is_cruncher_v3"`).
+
+- new function `download_jdemetra()` to download the JDemetra+ software.
+
+- improve `cruncher()` function to handle the case when the `cruncher_bin_directory` parameter provided correspond to the path to the JWSACruncher instead of the bin directory.
+
+- improve error message of of `create_param_file()` when the `dir_file_param` parameter doesn't correspond to an existing directory.
+
+- new function `check_cruncher_bin_directory()` to check if the `cruncher_bin_directory` parameter provided correspond to an existing directory and contains the JWSACruncher executable.
+
 # rjwsacruncher 0.2.2
 
 - add new parameter `short_column_headers` to `create_param_file()`.